+++ jenkins/jenkins_jobs.groovy
@@ -1,3 +1,0 @@
-jobFactory.pluginJobs({
-requiredJDK "JAVA_11"
-})